fix: move blending desc to module
